feat(overview): editable transcripts in overview table
About now it's possible to edit segments directly in the overview table.
The user just needs to click on a segment and an text editor appears. If
the user moves the mouse out of the text editor the segment is saved.
This will make it easier to correct small errors in transcripts.